requirements.txt
acres==0.5.0
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
ci-info==0.3.0
click==8.1.8
configobj==5.0.9
configparser==7.2.0
cryptography==45.0.5
distro==1.9.0
Django==4.2
django-admin-volt==1.0.17
django-ckeditor==6.7.3
django-js-asset==2.2.0
django-phonenumber-field==8.1.0
et_xmlfile==2.0.0
etelemetry==0.3.1
exceptiongroup==1.3.0
filelock==3.18.0
fitz==0.0.1.dev2
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
isodate==0.7.2
jiter==0.10.0
looseversion==1.3.0
lxml==6.0.0
networkx==3.2.1
nibabel==5.3.2
nipype==1.10.0
numpy==2.0.2
openai==1.99.7
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
pathlib==1.0.1
phonenumbers==9.0.10
pillow==11.3.0
prov==2.1.1
psycopg2-binary==2.9.10
puremagic==1.30
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydot==4.0.1
PyMuPDF==1.26.3
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyxnat==1.6.3
rdflib==7.1.4
requests==2.32.4
scipy==1.13.1
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
traits==7.0.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
zipp==3.23.0
